==> src/pipeline/s6_stats_data.ipynb <==
import soundfile as sf
from tqdm import tqdm
import pandas as pd
import librosa
import json
import os

from jiwer import wer

def load_cached_metadata(filepaths):
    metadata = []
    for filepath in filepaths:
        for line in tqdm(open(filepath).readlines()):
            sample = json.loads(line)
            metadata.append(sample)
            
    return metadata

filepaths = [
    "/data/asr-research/data/s9_stt_metadata_nemo_ctc/predictions_all.jsonl",
]
metadata = load_cached_metadata(filepaths)
metadata = pd.DataFrame(metadata)
metadata.head(1)

metadata["wer"] = metadata.apply(
    lambda row: wer(row["text"], row["pred_text"]), axis=1
)

subset = metadata[metadata.wer == 0]

metadata[metadata.wer == 0].duration.sum() / 3600

row = metadata[metadata.wer > 0.15].iloc[25].to_dict()

audio_filepath = row["audio_filepath"]
start_time = row["offset"]
end_time = start_time + row["duration"]

audio, sr = librosa.load(audio_filepath, sr=None)

audio = audio[int(sr*start_time): int(sr*end_time)]
sf.write("segment.wav", audio, samplerate=sr)

row

metadata.duration.sum()/ 3600

vocab = subset.text.apply(lambda x: x.split()).explode().value_counts()

outlier = vocab[vocab < 10]
outlier = set(outlier.index.to_list())

is_valid = subset.text.apply(lambda x: any([word in outlier for word in x.split()]))

subset[is_valid].duration.sum() / 3600

subset[~is_valid].duration.sum() / 3600

clean_data = subset[~is_valid]
clean_data = clean_data[
    [
        "audio_filepath", "offset", 
        "duration", "pred_text", "wer"
        ]
    ].rename(columns={"pred_text": "text"})

filepath = "data/f88_cleaned_data.jsonl"
with open(filepath, "w") as f:
    for index in tqdm(clean_data.index):
        row = clean_data.loc[index].to_dict()
        json_obj = json.dumps(row, ensure_ascii=False)
        f.write(json_obj+"\n")
print(f'saved data to {filepath}')
